Fix storybook display
This includes 2 pieces:

1. Moving all of the webpack customization into the .beaverrc file.
2. Update the webpack configuration for storybook to pull through
the .beaverrc.js file and extract those changes to be applied to the
storybook configuration.

Also add the missing decorator and fix the autoloader path so
everything continues to work as expected.
