+++ .storybook/main.js
@@ -1,27 +1,38 @@
-const {
-  lazyStyleRule,
-  styleLintPlugin,
-  notifierPlugin,
-  prismLanguageGenerationPlugin
-} = require('../webpack');
+const path = require('path');
+const { App, webpack } = require('brookjs-cli');
+
+const brookConfig = webpack.selectWebpackConfig({
+  cmd: 'start',
+  cwd: path.join(__dirname, '..'),
+  env: 'development',
+  extension: 'ts',
+  watch: false,
+  rc: App.create('beaver').getRC()
+});
 
 module.exports = {
   stories: ['../client/**/__stories__/*.stories.tsx'],
-  addons: [
-    '@storybook/addon-actions',
-    '@storybook/addon-links',
-    'brookjs-desalinate/register'
-  ],
-  webpack: async config => {
-    config.module.rules.splice(2, 1);
-    config.module.rules[0].test = /\.(tsx?|mjs|jsx?)$/;
-    config.module.rules.push(lazyStyleRule);
-    config.plugins.push(
-      styleLintPlugin,
-      notifierPlugin,
-      prismLanguageGenerationPlugin
-    );
-  
-    return config;
-  }
+  addons: ['@storybook/addon-actions', 'brookjs-desalinate/register'],
+  webpackFinal: async config => ({
+    ...config,
+    module: {
+      ...config.module,
+      rules: config.module.rules
+        // Remove the default CSS handler.
+        .filter(rule => !rule.test.source.includes('css'))
+        // Replace built-in JS babel-loader rule w/ our babel-loader.
+        .map(rule =>
+          rule.test.source.includes('js') ? brookConfig.module.rules[1] : rule
+        )
+        // Add our custom rules, including built-in styles + user modified.
+        .concat(...brookConfig.module.rules.slice(3))
+    },
+    // Make sure we resolve all extensions.
+    resolve: {
+      ...config.resolve,
+      extensions: brookConfig.resolve.extensions
+    },
+    // Add user custom plugins.
+    plugins: [...config.plugins, ...brookConfig.plugins.slice(5)]
+  })
 };
